% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gissid.R
\docType{data}
\name{gissid}
\alias{gissid}
\title{Great stellated dodecahedron}
\format{
A numeric matrix with 32 rows.
}
\usage{
gissid
}
\description{
The vertices of the great stellated dodecahedron (Bowers 
  acronym: gissid).
}
\keyword{datasets}
