% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-greatStellatedDodecahedron.R
\docType{data}
\name{greatStellatedDodecahedron}
\alias{greatStellatedDodecahedron}
\title{Great stellated dodecahedron}
\format{
A numeric matrix with 32 rows.
}
\usage{
greatStellatedDodecahedron
}
\description{
The vertices of the great stellated dodecahedron.
}
\keyword{datasets}
