% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullAS.R
\name{setAlpha}
\alias{setAlpha}
\title{Set alpha value to a full alpha hull and computes this alpha hull}
\usage{
setAlpha(ahull, alpha = "solid")
}
\arguments{
\item{ahull}{a full alpha hull, i.e. an output of \code{\link{fullAhull3d}}}

\item{alpha}{this xxx}
}
\value{
A \strong{rgl} mesh (class \code{mesh3d}) with the value of 
  alpha in the \code{"alpha"} attribute.
}
\description{
xx
}
\examples{
library(AlphaHull3D)
ahull <- fullAhull3d(greatStellatedDodecahedron)
mesh <- setAlpha(ahull, alpha = optimal(1))
library(rgl)
open3d(windowRect = c(50, 50, 562, 562))
shade3d(mesh, color = "maroon")
}
